src/evaluation/requirement.txt
absl-py
py-rouge==1.1
nltk
pandas
numpy
scikit-learn
scipy